refactor(gemma4_unified): split modules under 500-line cap
Extract the pure token-type/block-id derivation into src/vision/gemma4_unified_mask.rs and the unified loader+sanitize into src/loading/vlm_gemma_unified.rs (with their own test files), keeping gemma4_unified.rs and vlm_gemma.rs under the project's 500-line limit.
