+++ src/loading/vlm.rs
@@ -39,6 +39,8 @@ use models::sanitize_config_json;
 
 #[path = "vlm_gemma.rs"]
 mod gemma;
+#[path = "vlm_gemma_unified.rs"]
+mod gemma_unified;
 #[path = "vlm_internvl.rs"]
 mod internvl;
 #[path = "vlm_llava.rs"]
@@ -56,9 +58,8 @@ mod special;
 #[path = "vlm_youtu_vl.rs"]
 mod youtu_vl_loader;
 
-pub(crate) use gemma::{
-    load_gemma3_vlm, load_gemma3n_vlm, load_gemma4_unified, load_gemma4_vlm,
-};
+pub(crate) use gemma::{load_gemma3_vlm, load_gemma3n_vlm, load_gemma4_vlm};
+pub(crate) use gemma_unified::load_gemma4_unified;
 pub(crate) use internvl::load_internvl_vlm;
 pub(crate) use llava::{load_llava_bunny_vlm, load_llava_vlm};
 pub(crate) use nemotron_h_nano_omni::load_nemotron_h_nano_omni_vlm;
